Split ray maths types into multiple files in a folder, from multiple implementations in a file
