+++ src/application.cpp
@@ -4,6 +4,8 @@
 #include <chrono>
 #include <iostream>
 
+#include "raymath/raymath.hpp"
+
 
 RayTracerApplication::RayTracerApplication()
 {
