+++ src/raymath/raymath.hpp
@@ -1,0 +1,10 @@
+#ifndef _RAYTRACERRAYMATH_
+#define _RAYTRACERRAYMATH_
+
+
+#include "point.hpp"
+#include "direction.hpp"
+#include "ray.hpp"
+
+
+#endif
